+++ src/@core/Header/Desktop/components/Search/components/SearchButton.tsx
@@ -5,7 +5,7 @@ import { useSearchPriceStore } from "@core/Header/store/SearchPriceStore";
 import SearchIcon from "@mui/icons-material/Search";
 import { FetchSearchPost } from "@shared/components/FetchList/FetchList";
 import { SubletPostStore } from "@store/SubletPostStore";
-import * as hs from "./Header.styles";
+import * as hs from "../../Header.styles";
 
 const SearchButton = () => {
   const { searchDate } = useSearchDateStore();
