+++ src/@shared/components/RoomInfo/RoomReservation.tsx
@@ -1,7 +1,7 @@
 import * as s from "@shared/styles/Public.styles";
 import * as RS from "@shared/styles/RoomInfo.styles";
 
-import SearchDate from "@core/Header/Desktop/components/SearchDate";
+import SearchDate from "@core/Header/Desktop/components/Search/components/SearchDate";
 import { getDateDiff } from "@shared/components/StaticComponents/StaticComponents";
 
 import type { Post } from "@type/Type";
